+++ src/furnitures/entities/furniture.entity.ts
@@ -15,7 +15,7 @@ export class Furniture {
   drawer: number;
   @Column()
   fillingdrawer: number;
-  @Column({ type: 'float8' })
+  @Column({ type: 'decimal' })
   price: number;
   @Column()
   type: string;
